+++ test/test_helper.rb
@@ -22,4 +22,4 @@ ActiveRecord::Migration.create_table :users, force: true do |t|
   t.string :encrypted_email
 end
 
-Hypershield.quiet = false
+Hypershield.log_sql = true