+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.1.1 [unreleased]
 
 - Fixed error with MySQL 8
-- Added `quiet` option
+- Added `log_sql` option
 
 ## 0.1.0